+++ change-log.md
@@ -2,6 +2,7 @@
 
 ## 0.9
 - update dependencies to latest versions, including Selenium 3.0.1.
+- fix: OpenBrowser step should not propogate an exception during test resource cleanup if WebDriver.quit() throws an exception.
 
 ## 0.8
 - add: RandomNumberValueSource and random(min,max) expression support.